NIFI-5752: Load balancing fails with wildcard certs
NIFI-5752: Remove an unnecessary String.format

NIFI-5752: Remove an unnecessary block

This closes #3110.
