+++ nifi-nar-bundles/nifi-framework-bundle/nifi-framework/nifi-framework-core/src/main/java/org/apache/nifi/controller/queue/clustered/server/LoadBalanceAuthorizer.java
@@ -17,16 +17,17 @@
 
 package org.apache.nifi.controller.queue.clustered.server;
 
-import java.util.Collection;
+import javax.net.ssl.SSLSocket;
+import java.io.IOException;
 
 public interface LoadBalanceAuthorizer {
     /**
-     * Checks if any of the given identities is allowed to load balance data. If so, the identity that has been
-     * permitted is returned. If not, a NotAuthorizedException is thrown.
+     * Checks if the given SSLSocket (which includes identities) is allowed to load balance data. If so, the identity that has been
+     * permitted or hostname derived from the socket is returned. If not, a NotAuthorizedException is thrown.
      *
-     * @param clientIdentities the collection of identities to check
+     * @param sslSocket the SSLSocket which includes identities to check
      * @return the identity that is authorized, or null if the given collection of identities is null
      * @throws NotAuthorizedException if none of the given identities is authorized to load balance data
      */
-    String authorize(Collection<String> clientIdentities) throws NotAuthorizedException;
+    String authorize(SSLSocket sslSocket) throws NotAuthorizedException, IOException;
 }
